quick fix undo changes
